Fixes & improvements * Fix & improve blame parsing * Fix minimal hover box shown on empty commit * Fix hash action reolve remote address 1/2
Blame was not correct parsed. A filename and a summary was missed from the parsed blame info. Minimal hover box was shown for an empty commit. A check was added to prevent opening the hover box for an empty commit. HashAction did not open correctly the resolved remote address. Now resolve remote supports finding the git config from the workspace root or from the first sub applicable sub folder. This still does not support multiple git repositories under one workspace.